after_hours_voicemail/requirements.txt
google-genai>=1.14.0
telcoflow-sdk>=0.24.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
aws-sdk-bedrock-runtime>=0.1.0
smithy-aws-core>=0.1.0
python-dotenv>=1.1.0
